[feat]: #78 create command ping message
[body]

[notes]

close #78

+++ src/CommandHandler_Pong.cpp
@@ -8,7 +8,7 @@ void CommandHandler::checkPong(User &user) {
   std::time_t now = std::time(NULL);
   std::time_t diff = now - user.getPingTime();
 
-  if (diff < 10) {
+  if (static_cast<long>(diff) < 10) {
     return;
   }
   this->_server.delUser(user, "user Killed because of no respons\n\r");
